# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Delhi Ops Dashboard API
  version: 1.0.0
  description: Real-time urban monitoring dashboard for Delhi - Air Quality, Crime, Traffic, Cameras, and Incidents.
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    sortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    sortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    fromDateParam:
      name: from_date
      in: query
      schema:
        type: string
        format: date-time
    toDateParam:
      name: to_date
      in: query
      schema:
        type: string
        format: date-time
  schemas:
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    AqiReading:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        aqi:
          type: integer
          minimum: 0
          maximum: 999
        pm2_5:
          type: number
          nullable: true
        pm10:
          type: number
          nullable: true
        o3:
          type: number
          nullable: true
        no2:
          type: number
          nullable: true
        so2:
          type: number
          nullable: true
        co:
          type: number
          nullable: true
    CreateAqiRequest:
      type: object
      required: [location, aqi]
      properties:
        timestamp:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        aqi:
          type: integer
          minimum: 0
          maximum: 999
        pm2_5:
          type: number
          nullable: true
        pm10:
          type: number
          nullable: true
        o3:
          type: number
          nullable: true
        no2:
          type: number
          nullable: true
        so2:
          type: number
          nullable: true
        co:
          type: number
          nullable: true
    CrimeReport:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        severity:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [reported, investigating, resolved, closed]
        location:
          $ref: '#/components/schemas/Location'
    CreateCrimeRequest:
      type: object
      required: [type, location]
      properties:
        type:
          type: string
          maxLength: 100
        location:
          $ref: '#/components/schemas/Location'
        severity:
          type: string
          maxLength: 50
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [reported, investigating, resolved, closed]
...
    Camera:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [online, offline, maintenance]
        coordinates:
          type: string
        feed_url:
          type: string
          format: uri
        location:
          $ref: '#/components/schemas/Location'
    CreateCameraRequest:
      type: object
      required: [location, feed_url]
      properties:
        location:
          $ref: '#/components/schemas/Location'
        status:
          type: string
          enum: [online, offline, maintenance]
        feed_url:
          type: string
          format: uri
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: []
paths:
...
  /aqi:
    get:
      tags: [Air Quality]
      summary: List AQI readings
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/fromDateParam'
        - $ref: '#/components/parameters/toDateParam'
        - name: aqi_min
          in: query
          schema:
            type: integer
            minimum: 0
        - name: aqi_max
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Paginated AQI readings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AqiReading'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Air Quality]
      summary: Create AQI reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAqiRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AqiReading'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /crime:
    get:
      tags: [Crime]
      summary: List crime reports
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/fromDateParam'
        - $ref: '#/components/parameters/toDateParam'
        - name: type
          in: query
          schema:
            type: string
            maxLength: 100
        - name: severity
          in: query
          schema:
            type: string
            maxLength: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [reported, investigating, resolved, closed]
      responses:
        '200':
          description: Paginated crime reports
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CrimeReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Crime]
      summary: Create crime report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCrimeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/CrimeReport'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /cameras:
    get:
      tags: [Cameras]
      summary: List cameras
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortOrderParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, maintenance]
      responses:
        '200':
          description: Paginated cameras
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Camera'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Cameras]
      summary: Create camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCameraRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Camera'
        '400':
          $ref: '#/components/responses/BadRequest'
...